Move common functions to detailsUtils
https://issues.redhat.com/browse/COST-949
